asm: move terminator translation from local.go to terminator.go
