+++ asm/local.go
@@ -26,7 +26,6 @@ import (
 	"github.com/llir/l/ir/types"
 	"github.com/llir/l/ir/value"
 	"github.com/mewmew/l-tm/asm/ll/ast"
-	"github.com/mewmew/l-tm/internal/enc"
 	"github.com/pkg/errors"
 )
 
@@ -50,17 +49,7 @@ func newFuncGen(gen *generator, f *ir.Function) *funcGen {
 	}
 }
 
-// addLocal adds the local variable with the given name to the map of local
-// variables of the function.
-func (fgen *funcGen) addLocal(name string, v value.Value) error {
-	if prev, ok := fgen.ls[name]; ok {
-		return errors.Errorf("IR local identifier %q already present; prev `%s`, new `%s`", enc.Local(name), prev, v)
-	}
-	fgen.ls[name] = v
-	return nil
-}
-
-// resolveLocals resolves the local variables, basic blocks and function
+// resolveLocals resolves the local va1riables, basic blocks and function
 // parameters of the given function body. The returned value maps from local
 // identifier (without '%' prefix) to the corresponding IR value.
 func (fgen *funcGen) resolveLocals(body ast.FuncBody) (map[string]value.Value, error) {
@@ -83,7 +72,7 @@ func (fgen *funcGen) resolveLocals(body ast.FuncBody) (map[string]value.Value, e
 	// Translate terminators.
 	for i, block := range f.Blocks {
 		old := oldBlocks[i].Term()
-		if _, err := fgen.astToIRTerm(block.Term, old); err != nil {
+		if err := fgen.astToIRTerm(block.Term, old); err != nil {
 			return nil, errors.WithStack(err)
 		}
 	}
@@ -1491,192 +1480,6 @@ func (fgen *funcGen) astToIRInstCleanupPad(inst ir.Instruction, old *ast.Cleanup
 	}
 	// TODO: implement
 	return i, nil
-}
-
-// === [ Terminators ] =========================================================
-
-// astToIRTerm translates the AST terminator into an equivalent IR terminator.
-func (fgen *funcGen) astToIRTerm(term ir.Terminator, old ast.Terminator) (ir.Terminator, error) {
-	switch old := old.(type) {
-	// Value terminator.
-	case *ast.LocalDefTerm:
-		name := local(old.Name())
-		v, ok := fgen.ls[name]
-		if !ok {
-			return nil, errors.Errorf("unable to locate local variable %q", name)
-		}
-		t, ok := v.(ir.Terminator)
-		if !ok {
-			return nil, errors.Errorf("invalid terminator type of %q; expected ir.Terminator, got %T", name, v)
-		}
-		return fgen.astToIRValueTerm(t, old.Term())
-	case ast.ValueTerminator:
-		return fgen.astToIRValueTerm(term, old)
-	// Non-value terminators.
-	case *ast.RetTerm:
-		return fgen.astToIRTermRet(term, old)
-	case *ast.BrTerm:
-		return fgen.astToIRTermBr(term, old)
-	case *ast.CondBrTerm:
-		return fgen.astToIRTermCondBr(term, old)
-	case *ast.SwitchTerm:
-		return fgen.astToIRTermSwitch(term, old)
-	case *ast.IndirectBrTerm:
-		return fgen.astToIRTermIndirectBr(term, old)
-	case *ast.ResumeTerm:
-		return fgen.astToIRTermResume(term, old)
-	case *ast.CatchRetTerm:
-		return fgen.astToIRTermCatchRet(term, old)
-	case *ast.CleanupRetTerm:
-		return fgen.astToIRTermCleanupRet(term, old)
-	case *ast.UnreachableTerm:
-		return fgen.astToIRTermUnreachable(term, old)
-	default:
-		panic(fmt.Errorf("support for AST terminator type %T not yet implemented", old))
-	}
-}
-
-// astToIRValueTerm translates the AST value terminator into an equivalent IR
-// terminator.
-func (fgen *funcGen) astToIRValueTerm(term ir.Terminator, old ast.ValueTerminator) (ir.Terminator, error) {
-	switch old := old.(type) {
-	case *ast.InvokeTerm:
-		return fgen.astToIRTermInvoke(term, old)
-	case *ast.CatchSwitchTerm:
-		return fgen.astToIRTermCatchSwitch(term, old)
-	default:
-		panic(fmt.Errorf("support for value terminator %T not yet implemented", old))
-	}
-}
-
-// --- [ ret ] -----------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermRet(term ir.Terminator, old *ast.RetTerm) (*ir.TermRet, error) {
-	t, ok := term.(*ir.TermRet)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermRet), got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ br ] ------------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermBr(term ir.Terminator, old *ast.BrTerm) (*ir.TermBr, error) {
-	t, ok := term.(*ir.TermBr)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermBr, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-func (fgen *funcGen) astToIRTermCondBr(term ir.Terminator, old *ast.CondBrTerm) (*ir.TermCondBr, error) {
-	t, ok := term.(*ir.TermCondBr)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermCondBr, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ switch ] --------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermSwitch(term ir.Terminator, old *ast.SwitchTerm) (*ir.TermSwitch, error) {
-	t, ok := term.(*ir.TermSwitch)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermSwitch, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ indirectbr ] ----------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermIndirectBr(term ir.Terminator, old *ast.IndirectBrTerm) (*ir.TermIndirectBr, error) {
-	t, ok := term.(*ir.TermIndirectBr)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermIndirectBr, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ invoke ] --------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermInvoke(term ir.Terminator, old *ast.InvokeTerm) (*ir.TermInvoke, error) {
-	t, ok := term.(*ir.TermInvoke)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermInvoke, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ resume ] --------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermResume(term ir.Terminator, old *ast.ResumeTerm) (*ir.TermResume, error) {
-	t, ok := term.(*ir.TermResume)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermResume, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ catchswitch ] ---------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermCatchSwitch(term ir.Terminator, old *ast.CatchSwitchTerm) (*ir.TermCatchSwitch, error) {
-	t, ok := term.(*ir.TermCatchSwitch)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermCatchSwitch, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ catchret ] ------------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermCatchRet(term ir.Terminator, old *ast.CatchRetTerm) (*ir.TermCatchRet, error) {
-	t, ok := term.(*ir.TermCatchRet)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermCatchRet, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ cleanupret ] ----------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermCleanupRet(term ir.Terminator, old *ast.CleanupRetTerm) (*ir.TermCleanupRet, error) {
-	t, ok := term.(*ir.TermCleanupRet)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermCleanupRet, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
-}
-
-// --- [ unreachable ] ---------------------------------------------------------
-
-func (fgen *funcGen) astToIRTermUnreachable(term ir.Terminator, old *ast.UnreachableTerm) (*ir.TermUnreachable, error) {
-	t, ok := term.(*ir.TermUnreachable)
-	if !ok {
-		// NOTE: panic since this would indicate a bug in the implementation.
-		panic(fmt.Errorf("invalid IR terminator for AST terminator; expected *ir.TermUnreachable, got %T", term))
-	}
-	// TODO: implement
-	return t, nil
 }
 
 // ### [ Helper functions ] ####################################################
